requirements.txt
streamlit>=1.32.0
openai>=1.10.0
Pillow
